Make *UserController behavior customizable
